KVJI to KSIF
I hoped this would be from my RPi.  A revised version of only works on the RPi for 5 minutes.  I tried to modify the timeout but that did not work,  My Arduino BLE does not work right now either.  I completer this flight with an ancient version of gatttool on my linux desktop.